+++ src/main/java/edu/naukma/reshet/shared/DocumentaryFrequencyCrawler.java
@@ -1,5 +1,5 @@
 package edu.naukma.reshet.shared;
 
 public interface DocumentaryFrequencyCrawler {
-  public Long getDocumentaryFrequency(String term);
+  public Double getDocumentaryFrequency(String term);
 }
